+++ crates/nostr/src/event/id.rs
@@ -5,7 +5,6 @@
 //! Event Id
 
 use alloc::string::{String, ToString};
-use alloc::vec::Vec;
 use core::fmt;
 use core::str::FromStr;
 
@@ -111,8 +110,9 @@ impl EventId {
     where
         S: AsRef<[u8]>,
     {
-        let bytes: Vec<u8> = hex::decode(hex)?;
-        Self::from_slice(&bytes)
+        let mut bytes: [u8; 32] = [0u8; 32];
+        hex::decode_to_slice(hex, &mut bytes)?;
+        Ok(Self::owned(bytes))
     }
 
     /// Parse from bytes
@@ -235,5 +235,21 @@ mod tests {
         assert!(id.check_pow(16));
         assert!(id.check_pow(20));
         assert!(!id.check_pow(25));
+    }
+}
+
+#[cfg(bench)]
+mod benches {
+    use super::*;
+    use crate::prelude::hex::{decode, encode};
+    use crate::test::{black_box, Bencher};
+
+    const ID: &str = "2be17aa3031bdcb006f0fce80c146dea9c1c0268b0af2398bb673365c6444d45";
+
+    #[bench]
+    pub fn parse_event_id_from_hex(bh: &mut Bencher) {
+        bh.iter(|| {
+            black_box(EventId::from_hex(ID)).unwrap();
+        });
     }
 }